mgr/cephadm: move _refresh_host_daemons to serve.py
